feat(editingCard): add more granual button control

+++ src/resources/elements/editingCard/editingCard.ts
@@ -2,23 +2,38 @@ import { bindable } from "aurelia-framework";
 import "./editingCard.scss";
 
 export class EditingCard {
-  @bindable onDelete: () => void;
-  @bindable onEdit: () => void;
-  @bindable onSave: () => void;
+  @bindable onDelete: () => boolean | Promise<boolean>;
+  @bindable onEdit: () => boolean | Promise<boolean>;
+  @bindable onSave: () => boolean | Promise<boolean>;
 
   private viewMode: "edit" | "view" = "edit";
+  saving: boolean;
 
-  edit() {
+  async edit() {
+    if (typeof this.onEdit === "function") {
+      const canEdit = await this.onEdit();
+      if (!canEdit) return;
+    }
+
     this.viewMode = "edit";
-    this.onEdit();
   }
 
-  save() {
+  async save() {
+    this.saving = true;
+
+    if (typeof this.onSave === "function") {
+      const canSave = await this.onSave();
+      if (!canSave) return;
+    }
+
     this.viewMode = "view";
-    this.onSave();
+    this.saving = false;
   }
 
-  delete() {
-    this.onDelete();
+  async delete() {
+    if (typeof this.onDelete === "function") {
+      const canDelete = await this.onDelete();
+      if (!canDelete) return;
+    }
   }
 }
